+++ bindings/python/tree_sitter_t32/__init__.py
@@ -1,0 +1,5 @@
+"Lauterbach TRACE32 script language grammar for tree-sitter"
+
+from ._binding import language
+
+__all__ = ["language"]