+++ CHANGELOG.md
@@ -5,11 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.8.0] - 2024-03-23
+
+### Added
+
+- Language bindings for C, Go, Python, and Swift.
+- Switch to Tree-sitter v0.22
+
 ## [2.7.0] - 2024-03-03
 
 ### Added
 
-- Switch to Tree-sitter v0.21.0
+- Switch to Tree-sitter v0.21
 
 ### Changed